ai_app/pyproject.toml
[project]
name = "ai-app"
version = "0.1.0"
description = "AI Data Analysis Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair>=6.0.0",
    "fastapi>=0.135.2",
    "matplotlib>=3.10.8",
    "openai>=2.30.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_app", "ai_app.*"]
